server integration changes and fixes

+++ NMSReportingSuite/src/main/java/com/beehyv/nmsreporting/dao/MACourseAttemptDao.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface MACourseAttemptDao {
    List<FrontLineWorkers> getSuccessFulFirstCompletion(Date fromDate, Date toDate);
 
-    Date getFirstCompletionDate(Integer flwId);
+    Date getFirstCompletionDate(Long flwId);
 
     List<FrontLineWorkers> getSuccessFulCompletion(Date toDate);
 }
